feat: Add HyperEVM network integration
- Add custom HyperEVM chain definition with Chain ID 999
- Configure RPC endpoint: https://hyperliquid.drpc.org
- Add block explorer support: https://hyperevmscan.io
- Integrate HyperEVM into Wagmi provider chains
- Add HyperEVM support to success dialogs
- Organize custom chains in separate customChains.ts file
- Add factory address configuration for HyperEVM
- Enable HYPE token transactions and contract deployments

Network Details:
- Chain ID: 999 (0x3e7)
- Native Currency: HYPE (18 decimals)
- Mainnet: Active
- Factory Address: 0x7628d1fcf63BFCdB9d705fdB39B0C20de9B3f22E
