+++ app/providers.tsx
@@ -8,7 +8,7 @@ import {
 } from '@rainbow-me/rainbowkit'
 import { WagmiProvider } from 'wagmi'
 import { /* mainnet, polygon, optimism, */ arbitrum, /* base, zora, */ bscTestnet } from 'viem/chains'
-import { hyperEVM } from '@/config/networks'
+import { hyperEVM } from '@/config/customChains'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 
 if (!process.env.NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID) {
