+++ components/token-success-dialog.tsx
@@ -6,7 +6,7 @@ import { motion } from "framer-motion"
 import confetti from 'canvas-confetti'
 import { useChainId } from 'wagmi'
 import { type Chain, /* mainnet, polygon, optimism, */ arbitrum, /* base, zora, */ bscTestnet, /* bsc */ } from 'viem/chains'
-import { hyperEVM } from '@/config/networks'
+import { hyperEVM } from '@/config/customChains'
 import { ExternalLink } from 'lucide-react'
 
 interface TokenSuccessDialogProps {
