+++ components/vesting-success-dialog.tsx
@@ -5,7 +5,7 @@ import { motion } from "framer-motion"
 import confetti from 'canvas-confetti'
 import { useChainId } from 'wagmi'
 import { type Chain, arbitrum, bscTestnet } from 'viem/chains'
-import { hyperEVM } from '@/config/networks'
+import { hyperEVM } from '@/config/customChains'
 import { ExternalLink, Copy, Calendar, Users, Clock } from 'lucide-react'
 import { VestingFormValues } from "./vesting-creation-form"
 import { format } from "date-fns"
